requirements.txt
django
djangorestframework
djongo
pymongo==3.12.3
